@colyseus/auth: move responsibilities to their own separate files
