+++ packages/auth/src/index.ts
@@ -1,194 +1,15 @@
-import express from 'express';
-import grant, { GrantProvider, GrantConfig, GrantSession } from 'grant';
-import session from "express-session";
-import jsonwebtoken, { JwtPayload, Jwt } from "jsonwebtoken";
-import { expressjwt, Request } from 'express-jwt';
+import { JwtPayload, Jwt } from "jsonwebtoken";
+import { Request } from 'express-jwt';
 
-export type { Request, JwtPayload, Jwt };
+import { JsonWebToken } from "./JsonWebToken";
+import { auth } from './auth';
+import { OAuthCallback, oauth } from './oauth';
 
-export const JsonWebToken = {
-  options: {
-    /**
-     * The secret used to sign and verify the JWTs.
-     */
-    secret: process.env.JWT_SECRET as jsonwebtoken.Secret,
-
-    verify: {
-      /**
-       * The first algorithm in the list is used to sign new tokens.
-       */
-      algorithms: ['HS256'],
-    } as jsonwebtoken.VerifyOptions,
-  },
-
-  sign: function(payload: any, options: jsonwebtoken.SignOptions = {}) {
-    return new Promise<string>((resolve, reject) => {
-      if (options.algorithm === undefined) {
-        options.algorithm = JsonWebToken.options.verify.algorithms[0];
-      }
-
-      jsonwebtoken.sign(payload, this.options.secret, options, (err, token) => {
-        if (err) reject(err);
-        resolve(token);
-      });
-    });
-  },
-
-  verify: function (token: string, options: jsonwebtoken.VerifyOptions = this.options.verify) {
-    return new Promise<JwtPayload | Jwt | string>((resolve, reject) => {
-      jsonwebtoken.verify(token, this.options.secret, options, function (err, decoded) {
-        if (err) reject(err);
-        resolve(decoded);
-      });
-    });
-  },
-
-  /**
-   * Returns the decoded payload without verifying if the signature is valid
-   */
-  decode: jsonwebtoken.decode,
-
-  /**
-   * Get express middleware that verifies JsonWebTokens and sets `req.auth`.
-   */
-  middleware(params?: Partial<Parameters<typeof expressjwt>[0]>) {
-    if (!this.options.secret) {
-      console.error("❌ Please provide 'JWT_SECRET' environment variable, or set 'jwt.options.secret'.");
-    }
-
-    return expressjwt(Object.assign({
-      secret: this.options.secret,
-      // credentialsRequired: false,
-      algorithms: this.options.verify.algorithms,
-      ...this.options.verify,
-    }, params));
-  },
+export type {
+  Request,
+  JwtPayload,
+  Jwt,
+  OAuthCallback,
 };
 
-export type OAuthProviderName = '23andme' | '500px' | 'acton' | 'acuityscheduling' | 'adobe' | 'aha' | 'alchemer' | 'amazon' | 'angellist' | 'apple' | 'arcgis' | 'asana' | 'assembla' | 'atlassian' | 'auth0' | 'authentiq' | 'authing' | 'autodesk' | 'aweber' | 'axosoft' | 'baidu' | 'basecamp' | 'battlenet' | 'beatport' | 'bitbucket' | 'bitly' | 'box' | 'buffer' | 'campaignmonitor' | 'cas' | 'cheddar' | 'clio' | 'cognito' | 'coinbase' | 'concur' | 'constantcontact' | 'coursera' | 'crossid' | 'dailymotion' | 'deezer' | 'delivery' | 'deputy' | 'deviantart' | 'digitalocean' | 'discogs' | 'discord' | 'disqus' | 'docusign' | 'dribbble' | 'dropbox' | 'ebay' | 'echosign' | 'ecwid' | 'edmodo' | 'egnyte' | 'etsy' | 'eventbrite' | 'evernote' | 'eyeem' | 'facebook' | 'familysearch' | 'feedly' | 'figma' | 'fitbit' | 'flickr' | 'formstack' | 'foursquare' | 'freeagent' | 'freelancer' | 'freshbooks' | 'fusionauth' | 'garmin' | 'geeklist' | 'genius' | 'getbase' | 'getpocket' | 'gitbook' | 'github' | 'gitlab' | 'gitter' | 'goodreads' | 'google' | 'groove' | 'gumroad' | 'harvest' | 'hellosign' | 'heroku' | 'homeaway' | 'hootsuite' | 'huddle' | 'ibm' | 'iconfinder' | 'idme' | 'idonethis' | 'imgur' | 'infusionsoft' | 'instagram' | 'intuit' | 'jamendo' | 'jumplead' | 'kakao' | 'keycloak' | 'line' | 'linkedin' | 'live' | 'livechat' | 'logingov' | 'lyft' | 'mailchimp' | 'mailup' | 'mailxpert' | 'mapmyfitness' | 'mastodon' | 'medium' | 'meetup' | 'mendeley' | 'mention' | 'microsoft' | 'mixcloud' | 'moxtra' | 'myob' | 'naver' | 'nest' | 'netlify' | 'nokotime' | 'notion' | 'nylas' | 'okta' | 'onelogin' | 'openstreetmap' | 'optimizely' | 'osu' | 'patreon' | 'paypal' | 'phantauth' | 'pinterest' | 'plurk' | 'podio' | 'procore' | 'producthunt' | 'projectplace' | 'pushbullet' | 'qq' | 'ravelry' | 'redbooth' | 'reddit' | 'runkeeper' | 'salesforce' | 'sellsy' | 'shoeboxed' | 'shopify' | 'skyrock' | 'slack' | 'slice' | 'smartsheet' | 'smugmug' | 'snapchat' | 'snowflake' | 'socialpilot' | 'socrata' | 'soundcloud' | 'spotify' | 'square' | 'stackexchange' | 'stocktwits' | 'stormz' | 'storyblok' | 'strava' | 'stripe' | 'surveymonkey' | 'surveysparrow' | 'thingiverse' | 'ticketbud' | 'tiktok' | 'timelyapp' | 'todoist' | 'trakt' | 'traxo' | 'trello' | 'tripit' | 'trustpilot' | 'tumblr' | 'twitch' | 'twitter' | 'typeform' | 'uber' | 'unbounce' | 'underarmour' | 'unsplash' | 'untappd' | 'upwork' | 'uservoice' | 'vend' | 'venmo' | 'vercel' | 'verticalresponse' | 'viadeo' | 'vimeo' | 'visualstudio' | 'vk' | 'wechat' | 'weekdone' | 'weibo' | 'withings' | 'wordpress' | 'workos' | 'wrike' | 'xero' | 'xing' | 'yahoo' | 'yammer' | 'yandex' | 'zendesk' | 'zoom';
-export type OAuthProviderConfig = {
-  /**
-   * consumer_key or client_id of your OAuth app
-   */
-  key: string;
-
-  /**
-   * consumer_secret or client_secret of your OAuth app
-   */
-  secret: string;
-
-  /**
-   * array of OAuth scopes to request
-   */
-  scope?: string[];
-
-  /**
-   * generate random nonce string (OpenID Connect only)
-   */
-  nonce?: boolean;
-
-  /**
-   * custom authorization parameters
-   */
-  custom_params?: any;
-
-  /**
-   * relative route or absolute URL to receive the response data /hello | https://site.com/hey
-   */
-  callback?: string;
-
-  /**
-   * relative route or absolute URL to receive the response data /hello | https://site.com/hey
-   */
-  response?: Array<'tokens' | 'raw' | 'jwt' | 'profile'>
-}
-
-// TODO: automatically set up Redis if RedisDriver is being used.
-
-// import redis from "redis";
-// import connectRedis from "connect-redis";
-
-// const client = redis.createClient();
-// const RedisStore = connectRedis(session);
-
-type OAuthCallback = (data: GrantSession['response'], provider: OAuthProviderName) => Promise<unknown>;
-let oAuthCallback: (data: GrantSession['response'], provider: OAuthProviderName) => Promise<unknown> = async (data, provider) => {
-  console.debug("OAuth callback missing. Use oauth.onCallback() to persist user data.");
-  return data;
-};
-
-export const oauth = {
-  /**
-   * Default 'grant' module configuration.
-   */
-  defaults: {
-    // TODO: automatically retrieve "origin" based on "publicAddress" config [?]
-    // origin: process.env.BACKEND_URL,
-    transport: "session",
-    state: true,
-    response: ["tokens", "raw", "profile"],
-  } as GrantProvider & { prefix: never },
-
-  /**
-   * Route prefix for OAuth routes.
-   */
-  prefix: "/oauth",
-  providers: {} as { [providerId in OAuthProviderName]: OAuthProviderConfig },
-
-  /**
-   * Add OAuth provider configuration.
-   * @param providerId OAuth provider name
-   * @param config OAuth provider configuration
-   */
-  addProvider: function(providerId: OAuthProviderName, config: OAuthProviderConfig) {
-    this.providers[providerId] = config;
-  },
-
-  /**
-   * Provides a callback function that is called when OAuth is successful.
-   */
-  onCallback: function (callback: OAuthCallback) {
-    oAuthCallback = callback;
-  },
-
-  /**
-   * Returns an Express Router that handles OAuth for configured providers.
-   * @param callback (optional) Callback function that is called when OAuth is successful.
-   * @returns Express Router
-   */
-  callback: function (callback?: OAuthCallback) {
-    if (callback) {
-      this.onCallback(callback);
-    }
-
-    const router = express.Router();
-
-    const sessionMiddleware = session({
-      secret: process.env.SESSION_SECRET,
-      resave: false,
-      saveUninitialized: false, // true
-      // store: new RedisStore({ client: client }),
-    });
-
-    // set prefix
-    const config: GrantConfig = Object.assign({ defaults: this.defaults }, this.providers);
-    config.defaults.prefix = oauth.prefix;
-
-    router.use(sessionMiddleware);
-    router.use(grant.express(config));
-
-    router.get("/:providerId/callback", async (req, res) => {
-      const session = (req as any).session as unknown & { grant: GrantSession };
-      const data = await oAuthCallback(session.grant.response, session.grant.provider as OAuthProviderName);
-
-      /**
-       * I believe it is safe to use "*" in the origin here, since the token and origin are already validated by the OAuth provider.
-       * => https://developer.mozilla.org/en-US/docs/Web/API/Window/postMessage
-       */
-      res.send(`<!DOCTYPE html><html><head><script type="text/javascript">window.opener.postMessage(${JSON.stringify(data)}, '*');</script></head><body></body></html>`);
-      res.end();
-    });
-
-    return router;
-  }
-}
+export { JsonWebToken, auth, oauth, };
